+++ tests/e2e/assertions_test.go
@@ -11,10 +11,11 @@ import (
 
 func Test_E2E_Assertion_NoCalls(t *testing.T) {
 	state := e2eutils.NewState()
-	killMock, serverOutput, _ := e2eutils.RunMockBg(
+	killMock, serverOutput, _, _ := e2eutils.RunMockBg(
 		state,
 		"serve -c {{TEST_DATA_PATH}}/config_basic/config.json -p {{TEST_E2E_PORT}}",
 		nil,
+        true,
 	)
 
 	defer killMock()
@@ -38,10 +39,11 @@ func Test_E2E_Assertion_NoCalls(t *testing.T) {
 
 func Test_E2E_Assertion_BasicAssertion_WithValidationErrors(t *testing.T) {
 	state := e2eutils.NewState()
-	killMock, serverOutput, mockConfig := e2eutils.RunMockBg(
+	killMock, serverOutput, mockConfig, _ := e2eutils.RunMockBg(
 		state,
 		"serve -c {{TEST_DATA_PATH}}/config_basic/config.json -p {{TEST_E2E_PORT}}",
 		nil,
+        true,
 	)
 	defer killMock()
 
@@ -72,10 +74,11 @@ func Test_E2E_Assertion_BasicAssertion_WithValidationErrors(t *testing.T) {
 
 func Test_E2E_Assertion_WithNth(t *testing.T) {
 	state := e2eutils.NewState()
-	killMock, serverOutput, mockConfig := e2eutils.RunMockBg(
+	killMock, serverOutput, mockConfig, _ := e2eutils.RunMockBg(
 		state,
 		"serve -c {{TEST_DATA_PATH}}/config_basic/config.json -p {{TEST_E2E_PORT}}",
 		nil,
+        true,
 	)
 	defer killMock()
 
@@ -103,10 +106,11 @@ func Test_E2E_Assertion_WithNth(t *testing.T) {
 func Test_E2E_Assertion_WithNth_Failing(t *testing.T) {
 	state := e2eutils.NewState()
 
-	killMock, serverOutput, mockConfig := e2eutils.RunMockBg(
+	killMock, serverOutput, mockConfig, _ := e2eutils.RunMockBg(
 		state,
 		"serve -c {{TEST_DATA_PATH}}/config_basic/config.json -p {{TEST_E2E_PORT}}",
 		nil,
+        true,
 	)
 	defer killMock()
 
@@ -141,10 +145,11 @@ func Test_E2E_Assertion_WithNth_Failing(t *testing.T) {
 
 func Test_E2E_Assertion_WithNth_OutOfRange(t *testing.T) {
 	state := e2eutils.NewState()
-	killMock, serverOutput, mockConfig := e2eutils.RunMockBg(
+	killMock, serverOutput, mockConfig, _ := e2eutils.RunMockBg(
 		state,
 		"serve -c {{TEST_DATA_PATH}}/config_basic/config.json -p {{TEST_E2E_PORT}}",
 		nil,
+        true,
 	)
 	defer killMock()
 
@@ -175,10 +180,11 @@ func Test_E2E_Assertion_WithNth_OutOfRange(t *testing.T) {
 
 func Test_E2E_Assertion_BasicAssertion_WithoutValidationErrors(t *testing.T) {
 	state := e2eutils.NewState()
-	killMock, serverOutput, mockConfig := e2eutils.RunMockBg(
+	killMock, serverOutput, mockConfig, _ := e2eutils.RunMockBg(
 		state,
 		"serve -c {{TEST_DATA_PATH}}/config_basic/config.json -p {{TEST_E2E_PORT}}",
 		nil,
+        true,
 	)
 	defer killMock()
 
@@ -201,10 +207,11 @@ func Test_E2E_Assertion_BasicAssertion_WithoutValidationErrors(t *testing.T) {
 
 func Test_E2E_Assertion_Chaining_WithValidationErrors(t *testing.T) {
 	state := e2eutils.NewState()
-	killMock, serverOutput, mockConfig := e2eutils.RunMockBg(
+	killMock, serverOutput, mockConfig, _ := e2eutils.RunMockBg(
 		state,
 		"serve -c {{TEST_DATA_PATH}}/config_basic/config.json -p {{TEST_E2E_PORT}}",
 		nil,
+        true,
 	)
 	defer killMock()
 
@@ -240,10 +247,11 @@ func Test_E2E_Assertion_Chaining_WithValidationErrors(t *testing.T) {
 
 func Test_E2E_Assertion_Chaining_WithoutValidationErrors(t *testing.T) {
 	state := e2eutils.NewState()
-	killMock, serverOutput, mockConfig := e2eutils.RunMockBg(
+	killMock, serverOutput, mockConfig, _ := e2eutils.RunMockBg(
 		state,
 		"serve -c {{TEST_DATA_PATH}}/config_basic/config.json -p {{TEST_E2E_PORT}}",
 		nil,
+        true,
 	)
 	defer killMock()
 
@@ -274,10 +282,11 @@ func Test_E2E_Assertion_Chaining_WithoutValidationErrors(t *testing.T) {
 
 func Test_E2E_Assertion_MethodMatchingIsCaseInsensitive(t *testing.T) {
 	state := e2eutils.NewState()
-	killMock, serverOutput, mockConfig := e2eutils.RunMockBg(
+	killMock, serverOutput, mockConfig, _ := e2eutils.RunMockBg(
 		state,
 		"serve -c {{TEST_DATA_PATH}}/config_basic/config.json -p {{TEST_E2E_PORT}}",
 		nil,
+        true,
 	)
 	defer killMock()