+++ tests/e2e/base_api_edge_cases_test.go
@@ -17,7 +17,7 @@ func Test_E2E_BaseApi_RequestForwardedToBaseApi_SupportDifferentFormatsOfUrl(t *
 	}
 
 	state := NewState()
-	killMockBase, _, _ := RunMockBg(
+	killMockBase, _, _, _ := RunMockBg(
 		state,
 		strings.Join([]string{
 			"serve",
@@ -27,6 +27,7 @@ func Test_E2E_BaseApi_RequestForwardedToBaseApi_SupportDifferentFormatsOfUrl(t *
 			"--response 'Hello world! This is the base API.'",
 		}, " "),
 		nil,
+		true,
 	)
 	defer killMockBase()
 
@@ -44,6 +45,37 @@ func Test_E2E_BaseApi_RequestForwardedToBaseApi_SupportDifferentFormatsOfUrl(t *
 			strings.NewReader(""),
 			StatusCodeMatches(205),
 			StringMatches("Hello world! This is the base API."),
+		)
+	}
+}
+
+func Test_E2E_BaseApi_CannotBeStartedBecauseBaseApiIsInvalid(t *testing.T) {
+	baseApis := []string{
+		"....",
+		"x",
+		"--some-invalid-hostname--",
+		"http//invalid",
+		"http:invalid",
+		"http:///invalid",
+		"http://-----------.com",
+	}
+
+	for _, baseApi := range baseApis {
+		RunTestWithNoConfigAndWithArgsFailing(
+			t,
+			[]string{
+				fmt.Sprintf("--base %s", baseApi),
+				"--route hello/world",
+				"--response 'Hello world!'",
+			},
+			"GET",
+			"foo/bar",
+			nil,
+			strings.NewReader(""),
+			ApplicationOutputMatches([]string{
+				fmt.Sprintf("Base API is not valid: %s", baseApi),
+				"Set it as a valid domain name such as google.com",
+			}),
 		)
 	}
 }