+++ tests/e2e/delay_test.go
@@ -12,10 +12,11 @@ import (
 // The "delay" flag is not used here, therefore the request ends
 // quickly (in less than 2 seconds)
 func Test_E2E_Delay_WithoutDelay(t *testing.T) {
-	killMock, serverOutput, mockConfig := e2eutils.RunMockBg(
+	killMock, serverOutput, mockConfig, _ := e2eutils.RunMockBg(
 		e2eutils.NewState(),
 		"serve -c {{TEST_DATA_PATH}}/config_basic/config.json -p {{TEST_E2E_PORT}}",
 		nil,
+        true,
 	)
 	defer killMock()
 
@@ -35,10 +36,11 @@ func Test_E2E_Delay_WithoutDelay(t *testing.T) {
 
 // The "delay" flag is used, set to 3 seconds
 func Test_E2E_Delay_WithDelay(t *testing.T) {
-	killMock, serverOutput, mockConfig := e2eutils.RunMockBg(
+	killMock, serverOutput, mockConfig, _ := e2eutils.RunMockBg(
 		e2eutils.NewState(),
 		"serve -c {{TEST_DATA_PATH}}/config_basic/config.json -p {{TEST_E2E_PORT}} --delay 3000",
 		nil,
+        true,
 	)
 	defer killMock()