+++ tests/e2e/reset_test.go
@@ -11,10 +11,11 @@ import (
 
 func Test_E2E_Resetting(t *testing.T) {
 	state := e2eutils.NewState()
-	killMock, serverOutput, mockConfig := e2eutils.RunMockBg(
+	killMock, serverOutput, mockConfig, _ := e2eutils.RunMockBg(
 		state,
 		"serve -c {{TEST_DATA_PATH}}/config_basic/config.json -p {{TEST_E2E_PORT}}",
 		nil,
+        true,
 	)
 	defer killMock()